requirements.txt
# Investment Portfolio System - Requirements
# Python 3.12 compatible versions
#
# Note: Some optional ML/advanced features are commented out
# Install them later when needed:
#   - torch/transformers: For ML-based sentiment analysis
#   - ta-lib: For advanced technical indicators (requires C library)
#
# Core Dependencies
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Database & Time Series
timescale-vector==0.0.1
alembic==1.13.1

# Financial Data APIs
alpha-vantage==2.3.1
finnhub-python==2.4.19
fredapi==0.5.1
sec-edgar-downloader==5.0.2
pandas-datareader  # Stooq data source - free, unlimited historical data, no API key
anthropic          # Claude API for narrative analysis agents

# Data Analysis & ML
scikit-learn==1.4.0
scipy==1.12.0
statsmodels==0.14.1
# ta-lib==0.4.28  # Technical indicators (requires separate C library install - see note below)

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Web Dashboard
streamlit==1.30.0
dash==2.14.2
dash-bootstrap-components==1.5.0

# Task Scheduling
APScheduler==3.10.4
schedule==1.2.1

# Utilities
click==8.1.7  # CLI interface
tqdm==4.66.1  # Progress bars
loguru==0.7.2  # Better logging
pydantic==2.5.3  # Data validation
tenacity==8.2.3  # Retry logic

# Natural Language Processing (for sentiment - optional, install later if needed)
# transformers==4.37.0  # Uncomment when you need ML-based sentiment
# torch==2.5.1  # Updated for Python 3.12 compatibility

# API Development (future)
fastapi==0.109.0
uvicorn==0.27.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6
